Documentation/design/installconfig: Reword embedding subsection
I intially wanted to touch this line to fix the "controll" ->
"control" typo.  But I ended up rewording the subsection to lean more
obviously on our "we have opinions" position ;).  Of course, having
opinions doesn't mean we won't allow users to override those opinions.
I haven't gone into that level of detail, because we didn't have it
before, but I expect it has to do with "our operators are responsible
for cluster maintenance, and that's a lot more difficult as the number
of user knobs increases".